Fix a whole bunch of lurking bugs
Returning a newly constructed texture descriptor object in debug mode fails hard - the attached handles or object can become invalid when passed to another function. Using two new mutable variables, we can return a const reference to an internal class object that is updated if it is not ready to be used. This new class object references already existing objects that are also const-retrieved, so no more odd bugs!