This physical device logic.... a terrible idea
I did at least fix the stype traversal thing. That was horrid and broken beforehand. Problem is that we can't really do much with this, and I don't think we should've been anyways. Physical devices are a stupid object, as in literally dumb as bricks. Shouldn't do anything fancy with it really.